requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
black==22.10.0
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.4
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
telegram-send==0.34
tomli==2.0.1
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
